Removed int casting in rotate and added rotSpeed to Spaceship

+++ asteroids_local_backend/src/main/java/com/pluto/game/SpawnableEntity.java
@@ -63,9 +63,9 @@ public abstract class SpawnableEntity {
             localY = circle.position.y - position.y; // vector/circle class
 
             // Rotate local coordinates by radians
-            rotX = (int) Math.round( 
+            rotX = Math.round( 
                     localX * Math.cos(radians) - localY * Math.sin(radians));
-            rotY = (int) Math.round(
+            rotY = Math.round(
                     localX * Math.sin(radians) + localY * Math.cos(radians));
 
             // Convert back to global coordinates
